requirements.txt

gradio
matplotlib
torch
transformers
librosa
openai-whisper
